// src/interfaces/jmt.interface.ts
export interface combinScenes {
	name: string
	btName: string
	scenes: {
		scene: string
		thumbnail: string
	}[]
}